==> cm/code-01-analysis/step06_interpret_senses_subset.ipynb <==
import os
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import utilsCM
import csv


#Load word2sense
#already preprocessed in MATLAB so that Wrd2Sns and THINGs overlap --> we have IMAGES, LABELS and SENSES
pathtofile = '../code-00-preprocessdataset/'
W2S = pd.read_csv(pathtofile + "ThingsWrd2Sns.txt", sep=",")
Y_embeddings = W2S.values[:,1:W2S.shape[1]-1].astype(np.float) #just for threshold

tresh_bonf = utilsCM.p2r(.05/Y_embeddings.shape[1], 1470)

datapath = '../../../data-04/'
nsample = 12
WrdThingsInfo = pd.read_csv(datapath + 'KeptTHINGSInfo_n' + str(nsample) +'.csv',sep=',',index_col = 0)

# keyword = {'ROIpred'}
keyword = {'DNNActvtn','ROIpred'}

layer =  {'conv_1','conv_5','fc_3'}

ROI = {'EVC', 'ObjectROI'}

# Sub = [1,2,3,4]
Sub = [1,2,3,4]

#how many components to keep?
Keepncomps = [20]
pretrained_val = True

datapath = '../../../data-05/'
savepath = '../../../data-06/'

W2S = pd.read_csv(pathtofile + "ThingsWrd2Sns.txt", sep=",").set_index('item').drop(labels='sense2251', axis=1)
W2S_subset = W2S.iloc[WrdThingsInfo.old_index]

W2S_subset

##Showing histograms for senses above threshold
myDict = {}

for ilayer in layer:
    
    for ikeyword in keyword:        
        
        for icomps in Keepncomps:
            thisPrediction = []
            
            if ikeyword is 'DNNActvtn':
                if not pretrained_val:
                    filename = 'PredictSENSES_' + ikeyword + '_'+ ilayer + '_'+ str(icomps) +'PCs_untrained'
                    csvfilename = savepath + ikeyword+ "_" + ilayer + '_'+ str(icomps) +'PCs_untrained.csv'
                else:
                    filename = 'PredictSENSES_' + ikeyword + '_'+ ilayer + '_'+ str(icomps) +'PCs'
                    csvfilename = savepath + ikeyword+ "_" + ilayer + '_'+ str(icomps) +'PCs.csv'
                    
                thisPrediction = np.load(datapath + filename + '.npy')
                
                vecWithNan = np.argsort(thisPrediction)[::-1]#first 10 senses indeces, from highest to least hgih
                sortedIndeces = vecWithNan[~np.isnan(thisPrediction[vecWithNan])][0:9]
                
                Sample_W2S = W2S_subset.iloc[:,sortedIndeces]
                WeightedWords_rows = (Sample_W2S != 0).any(axis=1)#finds rows of zeros for those tops senses
                new_W2S = W2S_subset.loc[WeightedWords_rows] #takes them off from dataset
                
                ##Goes through each sense, prints out the top ten words and save them in a csv
                listOfWeights = []
                count = 0  
                for i in sortedIndeces:
                    count += 1
                    listOfWeights.append("Top "+ str(count) + " Sense " + str(i))
                    listOfWeights.append(new_W2S.iloc[:,i].sort_values(0)[::-1][0:9])
                    
                    
                
                #save in csv file
                with open(csvfilename, 'w', newline='') as csvfile:
                    wr = csv.writer(csvfile, quoting=csv.QUOTE_ALL)
                    wr.writerow(listOfWeights)
            
            
            elif ikeyword is 'ROIpred':
                
                for iROI in ROI:
                    thisPrediction = []
                    if not pretrained_val:
                        filename = 'PredictSENSES_' + ikeyword + '_' +iROI + '_'+ ilayer + '_'+ str(icomps) +'PCs_untrained'
                        csvfilename = savepath + ikeyword + "_" + iROI +'_' + ilayer + '_'+ str(icomps) +'PCs_untrained.csv'
                    else:
                        filename = 'PredictSENSES_' + ikeyword + '_' +iROI + '_'+ ilayer + '_'+ str(icomps) +'PCs'
                        csvfilename = savepath + ikeyword + "_" + iROI +'_' + ilayer + '_'+ str(icomps) +'PCs.csv'

                    
                    thisPrediction = np.load(datapath + filename + '.npy')
                    
                    vecWithNan = np.argsort(thisPrediction)[::-1] #first 10 senses indeces, from highest to least hgih
                    sortedIndeces = vecWithNan[~np.isnan(thisPrediction[vecWithNan])]
                    
                    Sample_W2S = W2S_subset.iloc[:,sortedIndeces]#eliminate rows of zeros
                    WeightedWords_rows = (Sample_W2S != 0).any(axis=1)
                    new_W2S = W2S_subset.loc[WeightedWords_rows] #takes them off from dataset

                    listOfWeights = []
                    count = 0
                    for i in sortedIndeces:
                        count += 1
                        listOfWeights.append("Top "+ str(count) + " Sense " + str(i))
                        listOfWeights.append(new_W2S.iloc[:,i].sort_values(0)[::-1][0:9])
                                    
                    with open(csvfilename, 'w', newline='') as csvfile:
                        wr = csv.writer(csvfile, quoting=csv.QUOTE_ALL)
                        wr.writerow(listOfWeights)


sortedIndeces

vec=np.argsort(thisPrediction)[::-1]
