==> cm/code-01-analysis/step07_SubsetSenses.ipynb <==
import os
import numpy as np
import pandas as pd
import csv


datapath = '../../../data-04/'
nsample = 12
WrdThingsInfo = pd.read_csv(datapath + 'KeptTHINGSInfo_n' + str(nsample) +'.csv',sep=',',index_col = 0)

#Load word2sense
#already preprocessed in MATLAB so that Wrd2Sns and THINGs overlap --> we have IMAGES, LABELS and SENSES
pathtofile = '../code-00-preprocessdataset/'
Y_embeddings = pd.read_csv(pathtofile + "ThingsWrd2Sns.txt", sep=",",index_col=0)
Y_embeddings

Y_embeddings_np = Y_embeddings.values[:,0:Y_embeddings.shape[1]-1].astype(np.float)
Y_embeddings_subset = Y_embeddings_np[WrdThingsInfo.old_index]
Y_embeddings_subset.shape #Pick only word subset of 312 objects

Y_embeddings

#Only select non-zero elements
#find indeced where they are all zeros
idx = np.argwhere(np.all(Y_embeddings_subset[:, :] == 0, axis=0) )
idx.shape

#retrieve name of the kept senses
idx_kept = np.argwhere(~np.all(Y_embeddings_subset[:, :] == 0, axis=0) )
idx_kept.shape
col_vec = []
col_vec = ["sense%03d" % (i+1) for i in idx_kept]

Y_embeddings_subset_nozero = np.delete(Y_embeddings_subset, idx, axis=1)
Y_embeddings_subset_nozero.shape

Dataset_subset = pd.DataFrame(Y_embeddings_subset_nozero,index = WrdThingsInfo['Word'],columns=col_vec)

savefilename = 'ThingsWrd2Sns_subset.txt'
savepath = '../../../data-07/'
#only keep senses with at least 10% of data is nonzero
vec = np.count_nonzero(Y_embeddings_subset_nozero, axis=0)

tenPercent = pd.array(vec>(Y_embeddings_subset_nozero.shape[0]*0.1),dtype='boolean')

Dataset_subset
Y_embeddings_subset_nozero_mostweight = Dataset_subset.iloc[:,tenPercent]
Y_embeddings_subset_nozero_mostweight.to_csv(savepath + savefilename)

savefilename

Y_embeddings_subset_nozero_mostweight
