==> cm/code-01-analysis/step10_preprocessWord2Vec.ipynb <==
import gensim
import pandas as pd
import numpy as np

model = gensim.models.KeyedVectors.load_word2vec_format('../../../../Word2Vec-GoogleNews/GoogleNews-vectors-negative300.bin.gz', binary=True)


dog = model['dog']
print(dog.shape)
print(dog[:10])

print(model.most_similar(positive=['woman', 'king'], negative=['man']))


print(model.doesnt_match("breakfast cereal dinner lunch".split()))


print(model.similarity('woman', 'man'))


datapath = '../../../data-04/'
nsample = 12
WrdThingsInfo = pd.read_csv(datapath + 'KeptTHINGSInfo_n' + str(nsample) +'.csv',sep=',',index_col = 0)

labels = WrdThingsInfo.Word.to_list()

# Deal with an out of dictionary word: Михаил (Michail)
# arr = [[] for _ in range(len(labels))]

for index,i  in enumerate(labels):
    if i in model:
        arr[index] = model[i]
    else:
        print('{0} is an out of dictionary word'.format(i))


W2vec = pd.DataFrame( np.array(arr) , index=labels)


savefilename = 'ThingsWrd2Vec_subset.txt'
savepath = '../../../data-10/'
W2vec.to_csv(savepath + savefilename)
