==> cm/code-01-analysis/step14_extract_features_things.ipynb <==
import os.path
import sys
sys.path.append('/Users/cmagri1/OneDrive - Johns Hopkins/git')
from torch import nn
import torch
from torchvision.models.alexnet import alexnet
from tqdm import tqdm

from EncodingModel_cm296 import utils as emutils
from EncodingModel_cm296 import feature_extractor as emfe

import pandas as pd
import numpy as np
import csv

from utilsCM import Subject
from utilsCM import cv_regression #Predicting fMRI reponses with features (Alexnet activations) 
#cv_regression computes cross-validated ridge regression. cross-validation groupings were pre-set based
#on fMRI design. 9-categories out, 9 folds. r is averaged over folds, weights are computed over all data


#Load word2sense
#already preprocessed in MATLAB so that Wrd2Sns and THINGs overlap --> we have IMAGES, LABELS and SENSES
pathtofile = '../code-00-preprocessdataset/'
Wrd2Sense = pd.read_csv(pathtofile + "ThingsWrd2Sns.txt", sep=",")
ImgInfo = pd.read_csv(pathtofile + "KeptTHINGSInfo.txt", sep=",")
pathtoTHINGS = '/Users/cmagri1/OneDrive - Johns Hopkins/Project-Word2Sense/THINGSdataset/Main/images'
PathToImgs = 'data-object2vec/stimuli'
savepath = '../../../data-14/'

#Only the subset
datapath = '../../../data-04/'
nsample = 12
WrdThingsInfo = pd.read_csv(datapath + 'KeptTHINGSInfo_n' + str(nsample) +'.csv',sep=',',index_col = 0)
WordInfThings = WrdThingsInfo.Word.to_numpy()

#Compute DNN features
def object_by_feature(PathToImgs,keyword,ilayer,resolutionval,cropval, savepath,WordInf=[],pretrained=True):
    if pretrainedModel:
        namefile = savepath+ keyword + '_' + ilayer 
    else:
        namefile = savepath+ keyword + '_' + ilayer +"_untrained"
        
    if os.path.isfile(namefile+".csv"): #if it's already saved
        
        print('loading file with '+ keyword+ ' stimuli features for ' + ilayer)
        features = pd.read_csv(namefile+".csv"  , sep=",", header=None, index_col=0)
        condition_features = {}
        for index, row in features.iterrows():
            condition_features[index] = row 
            
            
    else:
        
        if keyword == 'things':
            WordInf = WordInfThings
        
        
        print('Computing '+ keyword+ ' stimuli features for ' + ilayer)
        #Specify the model
        model = emfe.AlexNet(ilayer,pretrained_val=pretrained);      
        condition_features = emutils.compute_features(model, PathToImgs,resolutionval,cropval,WordInf)
#         features = 
        pd.DataFrame(condition_features).transpose().to_csv(namefile+".csv", index = True, header=False)
#         features.to_csv(namefile+".csv", index = True, header=False)
        np.save(namefile,condition_features)
#     layer_by_features[ilayer] = features; #create multistructure with all layers
    return condition_features

#Parameters
pretrainedModel = False
resolutionval = 256;
cropval = 224;


layer = {'conv_5'};
ROI = {'PFS'}

Sub = [1,2,3,4]

for ilayer in layer:    

     #Compute features of CNN
     #---------Object2Vect-----------------
    keyword = 'object2vec'
    object2vec_features = object_by_feature(PathToImgs,keyword,ilayer,resolutionval,cropval,savepath,pretrained=pretrainedModel)
    
#     #---------THINGS-----------------
    keyword = 'things'
    things_features = object_by_feature(pathtoTHINGS,keyword,ilayer,resolutionval,cropval,savepath,pretrained=pretrainedModel)
    things_features_df = pd.DataFrame(things_features).transpose()
    
    for iSub in Sub:
        for iROI in ROI:
            subject = Subject(iSub,[iROI])
            #predict fMRI using object2vec DNN activations
            weights, r = cv_regression(object2vec_features, subject, l2=0)
            if pretrainedModel:
                np.save(savepath + '/Weights_Sub' + str(iSub) + '_' + iROI + "_" + ilayer , weights)
            else:
                np.save(savepath + '/Weights_Sub' + str(iSub) + '_' + iROI + "_" + ilayer + '_untrained', weights)
                
            
            #     #---Compute Predicted ROI response to THINGS dataset
            print("Computing ROI " + iROI + " prediction for " + ilayer)    
            ROIpred = {}
            ROIpred = np.matmul(things_features_df.iloc[:,:].to_numpy(),weights.transpose())
            
            if pretrainedModel:
                np.save(savepath + '/ROIpred_Sub' + str(iSub) + '_' + iROI + "_" + ilayer, ROIpred)
            else:
                np.save(savepath + '/ROIpred_Sub' + str(iSub) + '_' + iROI + "_" + ilayer + '_untrained', ROIpred)
print('things dataset activations in DNN '+ ilayer +' size: ' + str(things_features_df.shape))
print('weights size: ' + str(weights.shape))
print('ROI pred size: ' + str(ROIpred.shape))
print("Done!")

# do it for random weights - permutation

import utilsCM
import matplotlib.pyplot as plt
# #Parameters
pretrainedModel = True
resolutionval = 227;


layer = {'conv_5'};
ROI = {'PFS'}
Sub = [1,2,3,4]

iters = 500
Randomstep = 'covMatching' #'permuteFeats' #covMatching

for ilayer in layer:    

     #Compute features of CNN
     #---------Object2Vect-----------------
    keyword = 'object2vec'
    object2vec_features = object_by_feature(PathToImgs,keyword,ilayer,resolutionval,cropval,savepath,pretrained=pretrainedModel)
    
#     #---------THINGS-----------------
    keyword = 'things'
    things_features = object_by_feature(pathtoTHINGS,keyword,ilayer,resolutionval,cropval,savepath,pretrained=pretrainedModel)
    things_features_df = pd.DataFrame(things_features).transpose()
    
    for iSub in Sub:
        for iROI in ROI:
            subject = Subject(iSub,[iROI])
            #predict fMRI using object2vec DNN activations
            print('computing Cv regression')
            weights, r = cv_regression(object2vec_features, subject, l2=0)
            savefigname = keyword  + '_Sub'+str(iSub)+ '_' +iROI + '_'+ ilayer
            plt.imshow(weights[:,0:250])
            plt.title(savefigname)
            plt.savefig(savepath + savefigname + '.png') 
            plt.close("all")
            
            
            #Do the random step
            if Randomstep is 'covMatching':
                print('computing randomMatrix')
                weights = utilsCM.RandomCovMatrix(weights)
                savefigname = keyword  + '_Sub'+str(iSub)+ '_' +iROI + '_'+ ilayer + 'random'
                plt.imshow(weights[:,0:250])
                plt.title(savefigname)
                plt.savefig(savepath + savefigname + '.png') 
                plt.close("all")
                
                np.save(savepath + '/Weights_Sub' + str(iSub) + '_' + iROI + "_" + ilayer + '_' + Randomstep, weights)
                print('computed randomMatrix')
#             if ilayer is not 'conv_1':
                savefigname = keyword + '_Sub'+str(iSub)+'_' +iROI + '_'+ ilayer + '_random'
                plt.imshow(weights[:,0:250])
                plt.title(savefigname)
                plt.savefig(savepath + savefigname + '.png') 
                plt.close("all")
                
                #     #---Compute Predicted ROI response to THINGS dataset
                print("Computing ROI " + iROI + " prediction for " + ilayer)    
                ROIpred = {}
                ROIpred = np.matmul(things_features_df.iloc[:,:].to_numpy(),weights.transpose())
                
                
            elif Randomstep is 'permuteFeats':
                print('permuting Features')
                ROIpred = utilsCM.PermuteFeats(weights,things_features_df.iloc[:,:].to_numpy(), iters)
            
            
            #save ROIpred
            np.save(savepath + '/ROIpred_Sub' + str(iSub) + '_' + iROI + "_" + ilayer + '_' + Randomstep, ROIpred)
            
print('things dataset activations in DNN '+ ilayer +' size: ' + str(things_features_df.shape))
print('weights size: ' + str(weights.shape))
print('ROI pred size: ' + str(ROIpred.shape))
print("Done!")

import numpy as np
x = [-2.1, -1,  4.3]
y = [3,  1.1,  0.12]
X = np.stack((x, y), axis=0)
cov_ = np.cov(X, rowvar = False)
cov_

mean_ = X.mean(axis=0)
mean_

np.random.multivariate_normal(mean_, cov_, 1)
