==> cm/code-01-analysis/step15_ConcatenateROIs.ipynb <==
import os
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import utilsCM

from tqdm import tqdm

import numpy as np
Ypredict = 'Word2Sense'

if Ypredict is 'Word2Vec':
    ### Load Word2Vec subset
    filename = 'ThingsWrd2Vec_subset.txt'
    filepath = '../../../data-10/'
    Wrd2Vec = pd.read_csv(filepath + filename,sep=',',index_col = 0)
    Y_embeddings = Wrd2Vec.values[:,:].astype(np.float)
elif Ypredict is 'Word2Sense':
    ### Load Word2Sense subset
    pathtofile = '../../../data-07/'
    Y_embeddings = pd.read_csv(pathtofile + "ThingsWrd2Sns_subset.txt", sep=",",index_col = 0)
    Y_embeddings = Y_embeddings.values[:,:].astype(np.float)

layer =  {'conv_5'}

Sub = [1,2,3,4]


pretrained_val = True

datapath = '../../../data-14/'
savepath = '../../../data-15/'

#make object selective ROIs by adding together LOC and PFS
objectROI = {'LOC','PFS'}
for ilayer in layer:
    for iSub in Sub:        
            
        if not pretrained_val:
            filename = datapath +  "ROIpred_Sub" + str(iSub) + '_ObjectROI_' + ilayer + '_untrained.npy'
        else:
            filename = datapath +  "ROIpred_Sub" + str(iSub) + '_ObjectROI_' + ilayer + '.npy'

        if not os.path.isfile(filename): #if it's not there
            predictor_variable = {}
            for iROI in objectROI:
                if not pretrained_val:
                    thisROI = np.load(datapath +  "ROIpred_Sub" + str(iSub) + '_' + iROI + "_" + ilayer + '_untrained.npy')
                else:
                    thisROI = np.load(datapath +  "ROIpred_Sub" + str(iSub) + '_' + iROI + "_" + ilayer + '.npy')
            
                #load ROIpred as predictor variable
                if len(predictor_variable) is 0:
                    predictor_variable = thisROI
                else:
                    predictor_variable = np.append( predictor_variable , thisROI, axis = 1)    
            np.save(filename,predictor_variable)
  

#make object selective ROIs by adding together LOC and PFS
objectROI = {'LOC','PFS'}
for ilayer in layer:
    for iSub in Sub:        
            
        filename = datapath +  "ROIpred_Sub" + str(iSub) + '_ObjectROI_' + ilayer + '_covMatching.npy'
       
        if not os.path.isfile(filename): #if it's not there
            predictor_variable = {}
            for iROI in objectROI:
                thisROI = np.load(datapath +  "ROIpred_Sub" + str(iSub) + '_' + iROI + "_" + ilayer + '_covMatching.npy')
            
                #load ROIpred as predictor variable
                if len(predictor_variable) is 0:
                    predictor_variable = thisROI
                else:
                    predictor_variable = np.append( predictor_variable , thisROI, axis = 1)    
            np.save(filename,predictor_variable)
  

#make scene selective ROIs by adding together OPA and OPA
objectROI = {'PPA','OPA'}
for ilayer in layer:
    for iSub in Sub:        
            
        if not pretrained_val:
            filename = datapath +  "ROIpred_Sub" + str(iSub) + '_SceneROI_' + ilayer + '_untrained.npy'
        else:
            filename = datapath +  "ROIpred_Sub" + str(iSub) + '_SceneROI_' + ilayer + '.npy'

        if not os.path.isfile(filename): #if it's not there
            predictor_variable = {}
            for iROI in objectROI:
                if not pretrained_val:
                    thisROI = np.load(datapath +  "ROIpred_Sub" + str(iSub) + '_' + iROI + "_" + ilayer + '_untrained.npy')
                else:
                    thisROI = np.load(datapath +  "ROIpred_Sub" + str(iSub) + '_' + iROI + "_" + ilayer + '.npy')
            
                #load ROIpred as predictor variable
                if len(predictor_variable) is 0:
                    predictor_variable = thisROI
                else:
                    predictor_variable = np.append( predictor_variable , thisROI, axis = 1)    
            np.save(filename,predictor_variable)
  

#make object selective ROIs by adding together LOC and PFS
objectROI = {'PPA','OPA'}
for ilayer in layer:
    for iSub in Sub:        
            
        filename = datapath +  "ROIpred_Sub" + str(iSub) + '_SceneROI_' + ilayer + '_covMatching.npy'
       
        if not os.path.isfile(filename): #if it's not there
            predictor_variable = {}
            for iROI in objectROI:
                thisROI = np.load(datapath +  "ROIpred_Sub" + str(iSub) + '_' + iROI + "_" + ilayer + '_covMatching.npy')
            
                #load ROIpred as predictor variable
                if len(predictor_variable) is 0:
                    predictor_variable = thisROI
                else:
                    predictor_variable = np.append( predictor_variable , thisROI, axis = 1)    
            np.save(filename,predictor_variable)
